Create README.md
This is a project that I did with my friend Zach. He expressed interest in potentially getting into Data Science, and I felt that it would be a good opportunity to brush up on my R skills while showing him the ropes. 

The data is from Kaggle, and the focus of the project is on data cleaning and visualizing.